升级时重置原有的本地数据库 每天随机从数据库选择推荐电影 天气预报显示优化 设置添加评分和豆瓣小组链接

+++ app/src/main/java/cn/sealiu/calendouer/bean/RatingBean.java
@@ -6,7 +6,7 @@ package cn.sealiu.calendouer.bean;
  */
 public class RatingBean {
     private int max;
-    private double average;
+    private float average;
     private String stars;
     private int min;
 
@@ -18,11 +18,11 @@ public class RatingBean {
         this.max = max;
     }
 
-    public double getAverage() {
+    public float getAverage() {
         return average;
     }
 
-    public void setAverage(double average) {
+    public void setAverage(float average) {
         this.average = average;
     }
